Move BlendFunctionFactors to Sprite and ParticleSystem only (#3900)
