+++ extensions/spine/Skeleton.js
@@ -29,7 +29,6 @@ const RenderComponent = require('../../cocos2d/core/components/CCRenderComponent
 const spine = require('./lib/spine');
 const SpineMaterial = require('../../cocos2d/core/renderer/render-engine').SpineMaterial;
 const Graphics = require('../../cocos2d/core/graphics/graphics');
-const BlendFactor = require('../../cocos2d/core/platform/CCMacro').BlendFactor;
 
 let SkeletonCache = require('./skeleton-cache');
 
@@ -113,44 +112,6 @@ sp.Skeleton = cc.Class({
          */
         paused: {
             default: false,
-            visible: false
-        },
-
-        /**
-         * !#en don't try to get or set srcBlendFactor,it doesn't affect,if you want to change spine blend mode,please set it in spine editor directly.
-         * !#zh 不要试图去获取或者设置 srcBlendFactor，没有意义，如果你想设置 spine 的 blendMode，直接在 spine 编辑器中设置即可。
-         * @property srcBlendFactor
-         * @type {macro.BlendFactor}
-         */
-        srcBlendFactor: {
-            get: function() {
-                return this._srcBlendFactor;
-            },
-            set: function(value) {
-                // shield set _srcBlendFactor
-            },
-            animatable: false,
-            type:BlendFactor,
-            override: true,
-            visible: false
-        },
-
-        /**
-         * !#en don't try to get or set dstBlendFactor,it doesn't affect,if you want to change spine blend mode,please set it in spine editor directly.
-         * !#zh 不要试图去获取或者设置 dstBlendFactor，没有意义，如果想设置 spine 的 blendMode，直接在 spine 编辑器中设置即可。
-         * @property dstBlendFactor
-         * @type {macro.BlendFactor}
-         */
-        dstBlendFactor: {
-            get: function() {
-                return this._dstBlendFactor;
-            },
-            set: function(value) {
-                // shield set _dstBlendFactor
-            },
-            animatable: false,
-            type: BlendFactor,
-            override: true,
             visible: false
         },
